+++ src/images/extendImageNode.ts
@@ -1,9 +1,9 @@
 import {GraphQLEnumType, GraphQLFieldConfigMap} from 'gatsby/graphql'
-import {PluginConfig} from '../gatsby-node'
 import {getCacheKey, CACHE_KEYS} from '../util/cache'
 import {ImageNode, ImageArgs, getGatsbyImageData} from './getGatsbyImageProps'
 
 import {getGatsbyImageFieldConfig} from 'gatsby-plugin-image/graphql-utils'
+import {PluginConfig} from '../util/validateConfig'
 
 const ImageFitType = new GraphQLEnumType({
   name: 'SanityImageFit',
@@ -44,7 +44,7 @@ export function extendImageNode(config: PluginConfig): GraphQLFieldConfigMap<any
 function getExtension(config: PluginConfig): GraphQLFieldConfigMap<any, any> {
   const location = {projectId: config.projectId, dataset: config.dataset}
   return {
-    gatsbyImageData: getGatsbyImageFieldConfig(
+    gatsbyImageData: getGatsbyImageFieldConfig<any, any, any>(
       (image: ImageNode, args: ImageArgs) => getGatsbyImageData(image, args, location),
       {
         placeholder: {
@@ -60,6 +60,6 @@ NONE: no placeholder.`,
           defaultValue: 'fill',
         },
       },
-    ),
+    ) as any,
   }
 }